% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_accueil.R
\name{mod_accueil_ui}
\alias{mod_accueil_ui}
\alias{mod_accueil_server}
\title{mod_accueil_ui and mod_accueil_server}
\usage{
mod_accueil_ui(id)

mod_accueil_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
