% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_chiffres_cles.R
\name{mod_chiffres_cles_ui}
\alias{mod_chiffres_cles_ui}
\alias{mod_chiffres_cles_server}
\title{mod_chiffres_cles_ui and mod_chiffres_cles_server}
\usage{
mod_chiffres_cles_ui(id)

mod_chiffres_cles_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
