% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_diplomes.R
\name{mod_diplomes_ui}
\alias{mod_diplomes_ui}
\alias{mod_diplomes_server}
\title{mod_diplomes_ui and mod_diplomes_server}
\usage{
mod_diplomes_ui(id)

mod_diplomes_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
