% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_emploi_30mois_employeur.R
\name{mod_emploi_30mois_employeur_ui}
\alias{mod_emploi_30mois_employeur_ui}
\alias{mod_emploi_30mois_employeur_server}
\title{mod_emploi_30mois_employeur_ui and mod_emploi_30mois_employeur_server}
\usage{
mod_emploi_30mois_employeur_ui(id)

mod_emploi_30mois_employeur_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
