% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_emploi_30mois_poste.R
\name{mod_emploi_30mois_poste_ui}
\alias{mod_emploi_30mois_poste_ui}
\alias{mod_emploi_30mois_poste_server}
\title{mod_emploi_30mois_poste_ui and mod_emploi_30mois_poste_server}
\usage{
mod_emploi_30mois_poste_ui(id)

mod_emploi_30mois_poste_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
