% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filtre_control_bar.R
\name{mod_filtre_control_bar_ui}
\alias{mod_filtre_control_bar_ui}
\alias{mod_filtre_control_bar_server}
\title{mod_filtre_control_bar_ui and mod_filtre_control_bar_server}
\usage{
mod_filtre_control_bar_ui(id)

mod_filtre_control_bar_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
