% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_poursuite_etudes.R
\name{mod_poursuite_etudes_ui}
\alias{mod_poursuite_etudes_ui}
\alias{mod_poursuite_etudes_server}
\title{mod_poursuite_etudes_ui and mod_poursuite_etudes_server}
\usage{
mod_poursuite_etudes_ui(id)

mod_poursuite_etudes_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
