% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_repondants.R
\name{mod_repondants_ui}
\alias{mod_repondants_ui}
\alias{mod_repondants_server}
\title{mod_repondants_ui and mod_repondants_server}
\usage{
mod_repondants_ui(id)

mod_repondants_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
