% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_temoignages.R
\name{mod_temoignages_ui}
\alias{mod_temoignages_ui}
\alias{mod_temoignages_server}
\title{mod_temoignages_ui and mod_temoignages_server}
\usage{
mod_temoignages_ui(id)

mod_temoignages_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
