% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_vie_active_durable.R
\name{mod_vie_active_durable_ui}
\alias{mod_vie_active_durable_ui}
\alias{mod_vie_active_durable_server}
\title{mod_vie_active_durable_ui and mod_vie_active_durable_server}
\usage{
mod_vie_active_durable_ui(id)

mod_vie_active_durable_server(input, output, session, rv)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
