% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(data, diplome, graph_font_family = NULL)
}
\arguments{
\item{data}{Table de données source}

\item{diplome}{Type de diplôme}

\item{graph_font_family}{Police de caractère utilisée}
}
\description{
Run the Shiny Application
}
